done with all positive and negative scenario testing

+++ src/main/java/com/cognologix/BankSystemApplicationAssignment/service/servicesimplementation/CustomerServicesImplementation.java
@@ -2,6 +2,10 @@ package com.cognologix.BankSystemApplicationAssignment.service.servicesimplement
 import com.cognologix.BankSystemApplicationAssignment.converter.Converter;
 import com.cognologix.BankSystemApplicationAssignment.dao.CustomerRepo;
 import com.cognologix.BankSystemApplicationAssignment.dto.CustomerDto;
+import com.cognologix.BankSystemApplicationAssignment.exceptions.AccountNotFoundException;
+import com.cognologix.BankSystemApplicationAssignment.exceptions.CustomerAlreadyExistException;
+import com.cognologix.BankSystemApplicationAssignment.exceptions.CustomerNotFoundException;
+import com.cognologix.BankSystemApplicationAssignment.model.Account;
 import com.cognologix.BankSystemApplicationAssignment.model.Customer;
 import com.cognologix.BankSystemApplicationAssignment.responses.CustomerResponse;
 import com.cognologix.BankSystemApplicationAssignment.service.serviceInterfaces.CustomerServices;
@@ -17,9 +21,6 @@ import java.util.stream.Collectors;
 
 @Service
 public class CustomerServicesImplementation implements CustomerServices {
-    public CustomerServicesImplementation(CustomerRepo customerRepo) {
-        this.customerRepo = customerRepo;
-    }
 
     @Autowired
     private CustomerRepo customerRepo;
@@ -27,26 +28,24 @@ public class CustomerServicesImplementation implements CustomerServices {
     private Converter converter;
     @Autowired
    private Converter customerConverter;
-    Random random=new Random();
     @Override
     public CustomerDto createNewCustomer(@Valid CustomerDto customerDto) {
         Customer customerDetails = this.customerConverter.customerDtoToModel(customerDto);
-        this.customerRepo.save(customerDetails);
-//        Account account=new Account();
-//        account.setCustomerId(customerDto.getCustomerId());
-//        account.setAccountNumber(account.getAccountNumber());
-//        account.setBankName(customerDto.getBankName());
-//        account.setAccountHolderName(customerDto.getCustomerName());
-//        account.setTypeOfAccount(customerDto.getTypeOfAccount());
-//        account.setTypeOfAccount(customerDto.getTypeOfAccount());
-//        account.setTotalAmount(0.0);
+        Optional<Optional<Customer>> customer= Optional.of(customerRepo.findById(customerDto.getCustomerId()));
+        if(customer.isPresent()){
+            throw new CustomerAlreadyExistException("Customer","CustomerId",customerDetails.getCustomerId());
+        }
         return this.customerConverter.customerModelToDto(customerDetails);
     }
     @Override
     public Optional<CustomerDto> getCustomerById(Integer customerId) {
-        Customer customer=this.customerRepo.findById(customerId).orElseThrow(()->new ResourceNotFoundException("Customer","Id",customerId));
-        return Optional.ofNullable(this.customerConverter.customerModelToDto(customer));
-
+        if(customerRepo.existsById(customerId)) {
+            Customer customer = this.customerRepo.findById(customerId).get();
+            return Optional.ofNullable(this.converter.customerModelToDto(customer));
+        }
+        else {
+            throw new CustomerNotFoundException("Customer","Id",customerId);
+        }
     }
     @Override
     public List<CustomerDto> findAllCustomerDetails() {
@@ -56,18 +55,28 @@ public class CustomerServicesImplementation implements CustomerServices {
         return cusDtos;
     }
     @Override
-    public CustomerDto updateCustomerDetails(CustomerDto customerDto) {
+    public CustomerResponse updateCustomerDetails(CustomerDto customerDto,Integer customerId) {
+        if(customerRepo.existsById(customerId)){
         Customer customerDetails = this.customerConverter.customerDtoToModel(customerDto);
-        //customerDetails.setCustomerId(customerId);
+        customerDetails.setCustomerId(customerId);
         this.customerRepo.save(customerDetails);
-        return this.customerConverter.customerModelToDto(customerDetails);
+        return new CustomerResponse("Customer details updated successfully",true);
+        }
+        else {
+            return new CustomerResponse("Customer Id does not exist",false);
+        }
 
     }
     @Override
     public CustomerResponse deleteCustomer(Integer customerId) {
-        this.customerRepo.deleteById(customerId);
-        CustomerResponse customerResponse=new CustomerResponse("Account deleted successfully..",true);
-        return customerResponse;
+        if(customerRepo.existsById(customerId)) {
+            this.customerRepo.deleteById(customerId);
+            CustomerResponse customerResponse = new CustomerResponse("Customer details deleted successfully..", true);
+            return customerResponse;
+        }
+        else {
+            return new CustomerResponse("Customer Id does not exist",false);
+        }
     }
 
 }
